+++ onlyoffice-monday-common/src/main/java/com/onlyoffice/common/tenant/transfer/event/AccessKeyRefreshed.java
@@ -10,6 +10,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AccessKeyRefreshed {
-  private int boardId;
-  private int tenantId;
+  private long boardId;
+  private long tenantId;
 }
