+++ onlyoffice-monday-common/src/main/java/com/onlyoffice/common/user/transfer/request/command/RemoveTenantUsers.java
@@ -14,5 +14,5 @@ import lombok.*;
 @AllArgsConstructor
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class RemoveTenantUsers {
-  @Positive private int tenantId;
+  @Positive private long tenantId;
 }
