+++ onlyoffice-monday-gateway/src/main/java/com/onlyoffice/gateway/processor/WebSocketNotificationProcessor.java
@@ -76,7 +76,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     }
   }
 
-  private boolean isSessionEligible(WebSocketSession session, int tenantId) {
+  private boolean isSessionEligible(WebSocketSession session, long tenantId) {
     if (!session.isOpen() || session.getUri() == null) return false;
     var token = getSessionToken(session);
     if (token == null) return false;
@@ -89,7 +89,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     return (tokenParams != null && !tokenParams.isEmpty()) ? tokenParams.getFirst() : null;
   }
 
-  private boolean verifyTenantId(String token, int tenantId) {
+  private boolean verifyTenantId(String token, long tenantId) {
     try {
       var claims = jwtDecoder.decode(token).getClaims();
       var sessionToken = mapper.convertValue(claims.get("dat"), SessionToken.class);
@@ -100,7 +100,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     }
   }
 
-  private void sendNotification(WebSocketSession session, TextMessage message, int tenantId) {
+  private void sendNotification(WebSocketSession session, TextMessage message, long tenantId) {
     try {
       MDC.put("tenant_id", String.valueOf(tenantId));
       MDC.put("session_id", session.getId());
