+++ onlyoffice-monday-tenant/src/main/java/com/onlyoffice/tenant/persistence/repository/BoardRepository.java
@@ -3,4 +3,4 @@ package com.onlyoffice.tenant.persistence.repository;
 import com.onlyoffice.tenant.persistence.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {}
+public interface BoardRepository extends JpaRepository<Board, Long> {}
