+++ onlyoffice-monday-tenant/src/main/java/com/onlyoffice/tenant/persistence/repository/TenantRepository.java
@@ -3,4 +3,4 @@ package com.onlyoffice.tenant.persistence.repository;
 import com.onlyoffice.tenant.persistence.entity.Tenant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantRepository extends JpaRepository<Tenant, Integer> {}
+public interface TenantRepository extends JpaRepository<Tenant, Long> {}
