+++ onlyoffice-monday-user/src/main/java/com/onlyoffice/user/persistence/entity/UserId.java
@@ -10,6 +10,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserId implements Serializable {
-  private int mondayId;
-  private int tenantId;
+  private long mondayId;
+  private long tenantId;
 }
